ctlsock: create exported ctlsock client library
The former interal ctlsock server package is renamed
to ctlsocksrv.
